feat(vscode): require explicit validation profile selection
Extract validation command, CLI lifecycle, and YAML schema contributor wiring from extension activation.

When no Azure DevOps profiles exist, validation now prompts to run offline or open profile management instead of silently falling back.

Refs #74, #80, #81.
