+++ docs/adr/0011-smart-validate-command-with-profile-picker.md
@@ -15,11 +15,13 @@ Three shapes were considered:
 
 ## Decision
 
-Use a single explicit `Atomize: Validate` command.
+Use a single explicit `Atomize: Validate` command. Validation Profile Selection is always user-driven — the system never chooses the validation mode silently.
 
 `Atomize: Validate` fetches Azure DevOps profiles from `atomize auth list --json`:
 - **One or more ADO profiles exist** → show a picker with those profiles plus an "Offline only" option.
-- **No ADO profiles exist** → run Offline Validation directly.
+- **No ADO profiles exist** → show a two-item picker: "Run offline" and "Add profile...". "Add profile..." awaits `manageProfiles` inline, then re-fetches profiles and re-enters Validation Profile Selection. This loops until the user picks a mode or dismisses the picker.
+
+The "Add profile..." item does not appear in the normal picker (one or more profiles present).
 
 The CLI default ADO profile remains the source of truth for connection-profile defaults. VS Code may mark or sort that profile first using the CLI's `isDefault` flag, but it must not run Online Validation silently. VS Code does not define a separate validation-default setting and does not offer to save a workspace validation default.
 
@@ -27,8 +29,9 @@ A single CLI invocation (`atomize validate --profile <name>`) produces the combi
 
 ## Consequences
 
-- Users with no ADO profiles experience no change to existing behavior.
+- Validation Profile Selection is shown on every run regardless of profile state.
 - Users with ADO profiles explicitly choose Online Validation or Offline Validation on each run.
+- Users with no ADO profiles are shown a path to add one rather than silently falling back to Offline Validation.
 - The CLI default ADO profile is visible in VS Code but not automatically selected.
 - There is no second VS Code default to become stale when a CLI profile is renamed or removed.
 - The panel title suffix `(Online)` / `(Offline)` is required to distinguish runs because the singleton panel persists after closing the file.
